% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregate,BSAExperiment-method}
\alias{aggregate,BSAExperiment-method}
\title{Aggregate Data in BSAExperiment}
\usage{
\S4method{aggregate}{BSAExperiment}(x, by, FUN, ...)
}
\arguments{
\item{x}{An object of class BSAExperiment.}

\item{by}{A formula indicating the grouping factors,
e.g., ~ condition + sac_day.}

\item{FUN}{Function to apply to each group for aggregation.}
}
\description{
Aggregates assay data in a BSAExperiment object based on specified
grouping factors in colData. The aggregation is performed by a
user-specified function, typically a sum, mean, etc.
}
