% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{applyFilter}
\alias{applyFilter}
\alias{applyFilter,BSAExperiment-method}
\title{Apply the variant, sample and/or matrix filters to a BSAExperiment object}
\usage{
\S4method{applyFilter}{BSAExperiment}(
  object,
  variant_filter = TRUE,
  sample_filter = TRUE,
  matrix_filter = TRUE
)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{variant_filter}{Boolean, default \code{TRUE}. Set to FALSE to use the
variant filter to remove variants from the BSAExperiment object.}

\item{sample_filter}{Boolean, default \code{TRUE}. Set to FALSE to use the
sample filter to remove samples from the BSAExperiment object.}

\item{matrix_filter}{Boolean, default \code{TRUE}. Set to FALSE to use the
matrix filter to remove entries from the BSAExperiment object.}
}
\value{
A BSAExperiment object filtered based on the supplied variant,
sample and/or entry filters
}
\description{
If any one of the selected, or by default, all, of the filters are set,
the corresponding entries in the BSAExperiment object will either be
removed, in the case of the variant or sample filter, or all entries
in the corresponding assays slots will be set to NA.
}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{comparisons<-}()},
\code{\link{deltaSnpIndex}()},
\code{\link{matrixFilter<-}()},
\code{\link{populationDepths}()},
\code{\link{refDepth}()},
\code{\link{sampleFilter<-}()},
\code{\link{snpIndex}()},
\code{\link{totalDepth<-}()},
\code{\link{variantFilter<-}()}
}
\concept{BSAExperiment-methods}
