% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BSAExperiment-class}
\alias{BSAExperiment-class}
\title{A class to store data related to bulk segregant analysis (BSA) experiments.}
\description{
BSAExperiment inherits from the \code{\link{RangedSummarizedExperiment}} class,
which is a subclass of the \code{\link{SummarizedExperiment}} class.
}
\details{
The \code{BSAExperiment} class
adds a \code{comparisons} slot to the \code{\link{RangedSummarizedExperiment}}
slots, which is a \code{DataFrame} object that describes comparisons between
two populations.
}
\section{Slots}{

\describe{
\item{\code{comparisons}}{A \code{DataFrame} object that describes comparisons between two
populations. The columns should be \code{population_1} and \code{population_2}, and
the levels of each should be a subset of the colData \code{sample} column.}
}}

\seealso{
\code{\link{RangedSummarizedExperiment}},
\code{\link{SummarizedExperiment}}
}
