% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{comparisons}
\alias{comparisons}
\alias{comparisons<-}
\alias{createComparisonsFrame}
\alias{populationDepths}
\alias{deltaAltFrequency}
\alias{comparisons,BSAExperiment-method}
\alias{comparisons<-,BSAExperiment-method}
\alias{createComparisonsFrame,BSAExperiment-method}
\alias{summary,BSAResults-method}
\alias{populationDepths,BSAExperiment-method}
\title{Get the BSAExperiment comparisons frame}
\usage{
comparisons(object, comparisons)

comparisons(object, ...) <- value

createComparisonsFrame(
  object,
  grouping_variable,
  population_variable,
  population_base_condition,
  var1_name,
  var2_name,
  base_cond_in_each_group
)

populationDepths(object, population_1_sample, population_2_sample)

deltaAltFrequency(object, population_1_sample, population_2_sample)

\S4method{comparisons}{BSAExperiment}(object)

\S4method{comparisons}{BSAExperiment}(object) <- value

\S4method{createComparisonsFrame}{BSAExperiment}(
  object,
  grouping_variable = "group",
  population_variable = "bulk",
  population_base_condition = "low",
  var1_name = "population_1",
  var2_name = "population_2",
  base_cond_in_each_group = TRUE
)

\S4method{summary}{BSAResults}(object)

\S4method{populationDepths}{BSAExperiment}(object, population_1_sample, population_2_sample)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{value}{A dataframe describing comparisons between two populations.
Must have two columns where the entries in each row are names of samples
in the \code{sample} column of the colData}

\item{grouping_variable}{column by which to group rows of the dataframe.
Defaults to 'batch'.}

\item{population_variable}{column which describes the various conditions
of each sample, eg if tissue, then the levels of this column might be
c(lung, brain, YPD, inoculum), Default: 'condition'}

\item{population_base_condition}{condition against which to compare all
other conditions, Default: 'inoculum'}

\item{var1_name}{the output frame will have two columns, the first storing
the samples which correspond to the \code{population_base_condition},
the other to the other sample conditions, Default: 'population_1'}

\item{var2_name}{as in var1_name, this will rename the second column
in the output frame, Default: 'population_2'}

\item{base_cond_in_each_group}{whether to include the base condition in
each group. Default TRUE}

\item{population_1_sample}{The name of the sample used as the first population.
In the context of the Takagi 2013 paper, this would be the 'low bulk' sample.}

\item{population_2_sample}{The name of the sample used for the second population.
In the context of the Takagi 2013 paper, this would be the 'high bulk' sample.}
}
\value{
a BSAExperiment object with the comparisons slot updated

A two column dataframe where the first column is the condition
against which all other sample conditions in that group are compared.
An example structure is:
\tabular{rcc}{ \tab population_2 \tab population_1 \cr \tab \eqn{P1.1I}
\tab \eqn{P1.1L} \cr \tab
\eqn{P1.1I} \tab \eqn{P1.1B} \cr}
}
\description{
Create a dataframe which describes comparisons between a given
condition and all other conditions described in a pool dataframe
}
\details{
This prepares samples for QTLseqr
}
\examples{
if (interactive()) {
  library(dplyr)
  # NOTE! "P2.1I" is a singleton
  sample_example <- c(
    "P1.1I", "P1.1L", "P1.1Y",
    "P1.2B", "P1.2I", "P1.2L", "P1.2Y", "P2.1I"
  )
  pool_construction <- tibble(sample = sample_example) \%>\%
    mutate(group = str_remove(sample, "\\\\w$")) \%>\%
    mutate(cond = ifelse(str_detect(sample, "P[[:alnum:]].{1,3}I"), "inoculum", NA)) \%>\%
    mutate(cond = ifelse(str_detect(sample, "P[[:alnum:]].{1,3}Y"), "ypd", cond)) \%>\%
    mutate(cond = ifelse(str_detect(sample, "P[[:alnum:]].{1,3}L"), "lung", cond)) \%>\%
    mutate(cond = ifelse(str_detect(sample, "P[[:alnum:]].{1,3}B"), "brain", cond)) \%>\%
    mutate(bulk = ifelse(cond == "inoculum", "low", "high"))
  sample_comparison_frame(pool_construction)
}

}
