% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{BSAExperiment}
\alias{BSAExperiment}
\title{A constructor for the BSAExperiment class}
\usage{
BSAExperiment(comparisons = DataFrame(), ...)
}
\arguments{
\item{comparisons}{Optional. By default an empty DataFrame(). Otherwise,
a dataframe describing comparisons between two populations.
Must have two columns where the entries in each row are names of samples
in the \code{sample} column of the colData.}

\item{...}{Additional arguments passed to \code{\link{SummarizedExperiment}}}
}
\description{
A constructor for the BSAExperiment class
}
\seealso{
\code{\link{SummarizedExperiment}}
}
