% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BSAResults-class}
\alias{BSAResults-class}
\title{Store statistical results of BSA Analysis}
\description{
The \code{BSAResults} class is a subclass of the \code{\link{RangedSummarizedExperiment}}
class, which is a subclass of the \code{\link{SummarizedExperiment}} class.
}
\details{
The \code{BSAResults} class is used to store the results of a BSA analysis.
}
\seealso{
\code{\link{BSAExperiment}},
\code{\link{RangedSummarizedExperiment}}
}
