% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGSmoothPvalues.R
\name{calculateGSmoothPvalues}
\alias{calculateGSmoothPvalues}
\title{Non-parametric estimation of the null distribution of G'}
\usage{
calculateGSmoothPvalues(g_smooth_vector, outlier_vector)
}
\arguments{
\item{g_smooth_vector}{a vector of smoothed G values (tricube weighted G statistics, sometimes called G')}

\item{outlier_vector}{is a boolean vector where if a value is TRUE, the corresponding
value in g_smooth_vector is considered an outlier}
}
\description{
Estimate p-values for the smoothed G statistic (sometimes called G') based
on the non-parametric estimation method described in Magwene et al. 2011.
The smoothed G statistic vector is filtered to exclude outlier regions based
on the \code{outlier_vector} -- see \code{\link{createBSAResults}} for more
details. The mode of the filtered G' set is estimated using the \code{\link[modeest]{mlv}}
with options \verb{bw = 0.5, method = 'hsm'}. Finally, the mean and variance
of the set are estimated using the median and mode and p-values
are estimated from a log normal distribution.
}
