% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{ChromQual}
\alias{ChromQual}
\title{ChromQual}
\usage{
ChromQual(
  vcf,
  chromlist = NULL,
  windowSize = NULL,
  ncol = NULL,
  HighLimQuality = NULL,
  Chromname = NULL,
  p1 = NULL,
  p2 = NULL,
  p3 = NULL,
  p4 = NULL,
  p5 = NULL,
  p6 = NULL,
  p7 = NULL
)
}
\arguments{
\item{vcf}{A vcf file}

\item{chromlist}{A vector specifying particular chromosomes}

\item{windowSize}{Specify window size to calculate number of SNPs}

\item{ncol}{An Integer Value Specifying the number of columns in ggplot facet_grid which corresponds to exact number of chromosomes in chromlist}

\item{HighLimQuality}{Set Upper Limit for Quality}

\item{p1}{TRUE or FALSE to plot or not to plot}

\item{p2}{TRUE or FALSE to plot or not to plot}

\item{p3}{TRUE or FALSE to plot or not to plot}

\item{p4}{TRUE or FALSE to plot or not to plot}

\item{p5}{TRUE or FALSE to plot or not to plot}

\item{p6}{TRUE or FALSE Boolean Argument, to plot or not to plot that is the question}

\item{p7}{TRUE or FALSE to plot or not}
}
\value{
Several ggplots
}
\description{
ChromQual
}
\examples{
ChromQuality(vcf = "General.vcf", chromlist = c("Chr1", "Chr2")), windowSize = 1e+06, scalar = 0.1, ncol = 2,HighLimQuality = 6000,  binwidth1 = 100, binwidth2 =1, DPBINS=10, p1=TRUE, p2=FALSE, p3=TRUE, p4=TRUE, p5=FALSE, p6=TRUE)
}
