% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{comparisons<-}
\alias{comparisons<-}
\alias{comparisons}
\alias{comparisons<-,BSAExperiment-method}
\alias{comparisons,BSAExperiment-method}
\title{Getter/Setter of the BSAExperiment comparison slot}
\usage{
\S4method{comparisons}{BSAExperiment}(object) <- value

\S4method{comparisons}{BSAExperiment}(object)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{value}{A dataframe describing comparisons between two populations.
Must have two columns where the entries in each row are names of samples
in the \code{sample} column of the colData}
}
\value{
The setter will return a BSAExperiment object with an updated
comparison slot while the getter will return the comparison slot's
DataFrame
}
\description{
Getter/Setter of the BSAExperiment comparison slot
}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{applyFilter}()},
\code{\link{deltaSnpIndex}()},
\code{\link{matrixFilter<-}()},
\code{\link{populationDepths}()},
\code{\link{refDepth}()},
\code{\link{sampleFilter<-}()},
\code{\link{snpIndex}()},
\code{\link{totalDepth<-}()},
\code{\link{variantFilter<-}()}
}
\concept{BSAExperiment-methods}
