\name{Correlation}
\alias{Correlation}
\title{Correlation}
\usage{
  Correlation(vcffile = "General.vcf", chromlist = NULL,p1 = NULL, p2 = NULL, p3 = NULL, p4 = NULL,p5=NULL) 
}
\arguments{
\item{vcffile}{A vcf file}
\item{chromlist}{A vector of Chromosome/Contig names reverse compatible with VCF file}
\item{p1}{Boolean TRUE or FALSE}
\item{p2}{Boolean TRUE or FALSE}
\item{p3}{Boolean TRUE or FALSE)
\item{p4}{Boolean TRUE or FALSE}
\item{p5}{Boolean TRUE or FALSE}
  }
  
}
\description{
  Returns correlation values, p-values, a heatmap, and correlogram 
}
