% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countSNPs_cpp}
\alias{countSNPs_cpp}
\title{Count number of SNPs within a sliding window}
\usage{
countSNPs_cpp(POS, windowSize)
}
\arguments{
\item{POS}{A numeric vector of genomic positions for each SNP}

\item{windowSize}{The required window size}
}
\description{
For each SNP returns how many SNPs are bracketing it within the set window size
}
