% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBSAResults.R
\name{createBSAResults}
\alias{createBSAResults}
\title{Calculate marker frequency statistics for a objectxperiment object}
\usage{
createBSAResults(
  object,
  comparison_index = NULL,
  statistics = c("delta_snp_index", "g"),
  ...
)
}
\arguments{
\item{object}{A BSAExperiment object}

\item{comparison_index}{a single integer value indicating the row of
comparisons(object) to utilize for the statistical comparison, or
a vector of integers indicating multiple comparisons to calculate}

\item{statistics}{A character vector indicating which statistics to calculate.
Currently 'delta_snp_index' and 'g' are valid choices}

\item{...}{
  Arguments passed on to \code{\link[=smoothAlleleFrequencyMetric]{smoothAlleleFrequencyMetric}}, \code{\link[=g_statistics_results]{g_statistics_results}}
  \describe{
    \item{\code{window_size}}{the window size (in base pairs) bracketing each allele for which
to calculate the statitics. Magwene et. al recommend a window size of ~25
cM, but also recommend optionally trying several window sizes to test if
peaks are over- or undersmoothed.}
  }}
}
\description{
Calculate marker frequency statistics for a objectxperiment object
}
