% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{createPopulationDepthList}
\alias{createPopulationDepthList}
\title{Population Depth List Object Constructor}
\usage{
createPopulationDepthList(depth_list)
}
\description{
Population Depth List Object Constructor
}
