% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBSAResults.R
\name{delta_snp_index_results}
\alias{delta_snp_index_results}
\title{Create a single comparison DataFrame}
\usage{
delta_snp_index_results(object, population_1_sample, population_2_sample, ...)
}
\arguments{
\item{object}{A objectxperiment object}

\item{population_1_sample}{One of the samples from object$sample which
will act as the "low bulk" or base sample for delta snp index calculation.}

\item{population_2_sample}{One of the samples from object$sample which
will act as the "high bulk" or comparison sample for delta snp index
calculation.}

\item{...}{Other arguments passed to \code{\link[locfit]{locfit}} and
subsequently to \code{\link[locfit]{locfit.raw}}() (or the lfproc). Usefull
in cases where you get "out of vertex space warnings"; Set the maxk higher
than the default 100. See \code{\link[locfit]{locfit.raw}}().}
}
\description{
This is an internal function which creates a DataFrame of QTLseqr results.
createBSAResult() should be used by general users as it returns the
BSAResults object which includes the rowRanges and colData. However, it
is exposed in order to provide documentation on the arguments which may
be passed to createBSAResults() in order to adjust how smoothing and
filtering is conducted, and what results are returned.
}
\keyword{internal}
