% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{deltaSnpIndex}
\alias{deltaSnpIndex}
\alias{deltaSnpIndex,BSAExperiment-method}
\title{Calculate \eqn{\Delta}{Delta}(Alt Allele Frequency) Between Two Populations}
\usage{
\S4method{deltaSnpIndex}{BSAExperiment}(object, population_1_sample, population_2_sample)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{population_1_sample}{The name of the sample used as the first
population. In the context of the Takagi 2013 paper, this would be the
'low bulk' sample.}

\item{population_2_sample}{The name of the sample used for the second
population. In the context of the Takagi 2013 paper, this would be the
'high bulk' sample.}
}
\value{
A numeric vector that results from
\eqn{Population_2\ Alternate\ Allele\ Frequency\ -\ Population_1\ Alternate\ Allele\ Frequency}
}
\description{
Calculate the difference in alternate allele frequency between population_2
and population_1
}
\examples{
# Assuming `bsae` is a BSAExperiment object with populations "Pop1"
# and "Pop2"
delta_freqs <- deltaSnpIndex(bsae, "Pop1", "Pop2")

# View the first few differences in allele frequency
head(delta_freqs)

}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{applyFilter}()},
\code{\link{comparisons<-}()},
\code{\link{matrixFilter<-}()},
\code{\link{populationDepths}()},
\code{\link{refDepth}()},
\code{\link{sampleFilter<-}()},
\code{\link{snpIndex}()},
\code{\link{totalDepth<-}()},
\code{\link{variantFilter<-}()}
}
\concept{BSAExperiment-methods}
