% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_hampel_filter.R
\name{.erfinv}
\alias{.erfinv}
\title{Complementary Error Function}
\usage{
.erfinv(y)
}
\description{
Complementary Error Function
}
\note{
copied from PRACMA, which is distributed under a GNU v3+
}
\author{
Copyright (c) 2015 Hans W Borchers
}
\keyword{internal}
