% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSAExperimentFromGatkTable.R
\name{.parse_ad_table}
\alias{.parse_ad_table}
\title{Parse AD Table}
\usage{
.parse_ad_table(ad_data)
}
\arguments{
\item{ad_data}{A matrix of AD data.}
}
\value{
A numeric matrix where each element is the second value from the
corresponding element in the original AD data.
}
\description{
This function takes a matrix of AD data, splits each element at the comma,
extracts the second value, converts these values to numeric, and reshapes
the result into a matrix with the same number of rows as the original AD data.
}
\examples{
ad_data <- matrix(c("1,149", "240,3", "225,65", "236,56", "406,141"), nrow = 5)
.parse_ad_table(ad_data)

}
\keyword{internal}
