% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSAExperimentFromGatkTable.R
\name{.parse_gt_table}
\alias{.parse_gt_table}
\title{Parse GT Table}
\usage{
.parse_gt_table(gt_data, ref)
}
\arguments{
\item{gt_data}{A data frame of GT data.}

\item{ref}{A character vector of reference values.}

\item{chunk_size}{An integer which sets the}
}
\value{
A numeric matrix where each element is 0 if the corresponding GT value
is equal to the reference and 1 otherwise.
}
\description{
This function takes a data frame of GT data and a reference column,
compares each GT value to the reference, and returns a numeric matrix
where each element is 0 if the corresponding GT value is equal to the reference
and 1 otherwise.
}
\examples{
gt_data <- data.frame(sample1 = c("A", "G", "C", "T", "A"),
                      sample2 = c("G", "A", "C", "T", "G"))
ref <- c("A", "G", "C", "T", "A")
.parse_gt_table(gt_data, ref)

}
\keyword{internal}
