% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGPrimeResults.R
\name{.validateCreateBSAResults}
\alias{.validateCreateBSAResults}
\title{Check the input values of createBSAResults}
\usage{
.validateCreateBSAResults(...)
}
\description{
Check the input values of createBSAResults
}
\keyword{internal}
