% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBSAResults.R
\name{.validateCreateBSAResultsArgs}
\alias{.validateCreateBSAResultsArgs}
\title{Check the input values of createBSAResults}
\usage{
.validateCreateBSAResultsArgs(args_list)
}
\description{
Check the input values of createBSAResults
}
\keyword{internal}
