% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_analysis.R
\name{extract_alt_freq_dataframe}
\alias{extract_alt_freq_dataframe}
\title{Extract alternative frequency data frame}
\usage{
extract_alt_freq_dataframe(
  bsae_sample_subset,
  depth_threshold = 10,
  ref_freq_thres = 0.2,
  hetero_freq_thres = c(0.4, 0.6),
  alt_freq_thres = 0.9,
  neighborhood_size = 20,
  verbose = FALSE,
  ncpus = 1,
  ...
)
}
\arguments{
\item{bsae_sample_subset}{A SummarizedExperiment object subset for a specific sample.}

\item{depth_threshold}{Integer specifying the minimum depth required for a
variant to be considered. Default is 10.}

\item{ref_freq_thres}{Numeric specifying the the maximum value for which
a given variant will be 'rounded' to REF (0) for the smoothed allele
frequency column}

\item{hetero_freq_thres}{Numeric vector specifying the range of values for
which a given variant will be 'rounded' to HET (0.5) for the smoothed
allele frequency column}

\item{alt_freq_thres}{Numeric specifying the minimum value for which a given
variant will be 'rounded' to ALT (1) for the smoothed allele frequency
column}

\item{neighborhood_size}{Integer specifying the number of data points around
a given variant that will be considered when determining if the variant
is 'noisy', or not matching its immediate neighbors. Default is 20.}

\item{verbose}{if TRUE, a message reporting the current sample will be
displayed If 2, both sample and chromosome messages will be displayed.}

\item{ncpus}{Number of cores to use for parallel processing.}

\item{...}{
  Arguments passed on to \code{\link[changepoint.np:cpt.np]{changepoint.np::cpt.np}}
  \describe{
    \item{\code{data}}{A vector, ts object or matrix containing the data within which you wish to find a changepoint.  If the data is a matrix, each row is considered as a separate dataset.}
    \item{\code{penalty}}{Choice of "None", "SIC", "BIC", "MBIC", AIC", "Hannan-Quinn", "Manual" and "CROPS" penalties.  If Manual is specified, the manual penalty is contained in the pen.value parameter. If CROPS is specified, the penalty range is contained in the pen.value parameter; note this is a vector of length 2 which contains the minimum and maximum penalty value.  Note CROPS can only be used if the method is "PELT". The predefined penalties listed DO count the changepoint as a parameter, postfix a 0 e.g."SIC0" to NOT count the changepoint as a parameter.}
    \item{\code{pen.value}}{The value of the penalty when using the Manual penalty option.  A vector of length 2 (min,max) if using the CROPS penalty.}
    \item{\code{method}}{Currently the only method is "PELT".}
    \item{\code{test.stat}}{The assumed test statistic/distribution of the data. Currently only "empirical_distribution".}
    \item{\code{class}}{Logical. If TRUE then an object of class cpt is returned.}
    \item{\code{minseglen}}{Positive integer giving the minimum segment length (number of observations between changes), default is the minimum allowed by theory.}
    \item{\code{nquantiles}}{The number of quantiles to calculate when test.stat = "empirical_distribution".}
  }}
}
\value{
A tibble containing the processed alternative frequency data.
}
\description{
Extract alternative frequency data frame
}
