% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{Facet_Allelic_Chrom}
\alias{Facet_Allelic_Chrom}
\title{Facet_Allelic_Chrom}
\usage{
Facet_Allelic_Chrom(
  SNPset,
  subset = NULL,
  var = "Allelicfreq",
  scaleChroms = TRUE,
  line = TRUE,
  thresholdGprime = NULL,
  thresholdHighAlternateFrequency = NULL
)
}
\arguments{
\item{SNPset}{A vcf file that has been imported by ImportVCF, filtered by SNPfilter, and run}

\item{subset}{A Chromosome subset}

\item{var}{This is a variable that you cannot change, it must be defined as Allelicfreq}

\item{scaleChroms}{TRUE or FASLE do you want plots on each Chromosome or NOT}

\item{line}{TRUE or FALSE do you want a line graph or just points}

\item{thresholdGprime}{Gprime Statistic cut off value}

\item{thresholdHighAlternateFrequency}{High Bulk Alternative Allele Frequency}
}
\description{
Provides plot of Alterante allelic depths for High and Low Bulk, sorry no sliding window.
}
