% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_args}
\alias{filter_args}
\title{Utility function to filter arguments}
\usage{
filter_args(FUN, ...)
}
\arguments{
\item{FUN}{a function}

\item{...}{arguments to extract from the ... args based on FUN}
}
\value{
a list of arguments that are valid for FUN
}
\description{
Utility function to filter arguments
}
