% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Filter.R
\name{filterSNPs}
\alias{filterSNPs}
\title{Filter SNPs based on read depth and quality}
\usage{
filterSNPs(SNPset, refAlleleFreq, filterAroundMedianDepth, minTotalDepth,
  maxTotalDepth, minSampleDepth, minGQ, verbose = TRUE)
}
\arguments{
\item{SNPset}{The data frame imported by \code{ImportFromGATK}}

\item{refAlleleFreq}{A numeric < 1. This will filter out SNPs with a
Reference Allele Frequency less than \code{refAlleleFreq} and greater than
1 - \code{refAlleleFreq}. Eg. \code{refAlleleFreq = 0.3} will keep SNPs
with 0.3 <= REF_FRQ <= 0.7}

\item{filterAroundMedianDepth}{Filters total SNP read depth for both bulks. A
median and median absolute deviation (MAD) of depth will be calculated.
SNPs with read depth greater or less than \code{filterAroundMedianDepth}
MADs away from the median will be filtered.}

\item{minTotalDepth}{The minimum total read depth for a SNP (counting both
bulks)}

\item{maxTotalDepth}{The maximum total read depth for a SNP (counting both
bulks)}

\item{minSampleDepth}{The minimum read depth for a SNP in each bulk}

\item{minGQ}{The minimum Genotype Quality as set by GATK. This is a measure
of how confident GATK was with the assigned genotype (i.e. homozygous ref,
heterozygous, homozygous alt). See
\href{http://gatkforums.broadinstitute.org/gatk/discussion/1268/what-is-a-vcf-and-how-should-i-interpret-it}{What
is a VCF and how should I interpret it?}}

\item{verbose}{logical. If \code{TRUE} will report number of SNPs filtered in
each step.}
}
\value{
Returns a subset of the data frame supplied which meets the filtering
  conditions applied by the selected parameters. If \code{verbose} is
  \code{TRUE} the function reports the number of SNPs filtered in each step
  as well as the initiatl number of SNPs, the total number of SNPs filtered
  and the remaining number.
}
\description{
Use filtering paramaters to filter out high and low depth reads as well as
low Genotype Quality as defined by GATK. All filters are optional but recommended.
}
\examples{
df_filt <- FilterSNPs(
    df,
    refAlleleFreq = 0.3,
    minTotalDepth = 40,
    maxTotalDepth = 80,
    minSampleDepth = 20,
    minGQ = 99,
    verbose = TRUE
)

}
\seealso{
See \code{\link[stats]{mad}} for explaination of calculation of
  median absolute deviation.
  \href{http://gatkforums.broadinstitute.org/gatk/discussion/1268/what-is-a-vcf-and-how-should-i-interpret-it}{What
  is a VCF and how should I interpret it?} for more information on GATK
  Fields and Genotype Fields
}
