% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_analysis.R
\name{find_breakpoints}
\alias{find_breakpoints}
\title{Find Breakpoints in Allele Frequency Data}
\usage{
find_breakpoints(df, ...)
}
\arguments{
\item{df}{A data frame containing smoothed allele frequencies and other
relevant data.}
}
\value{
A data frame indicating the positions of identified breakpoints and their characteristics.
}
\description{
Identifies breakpoints in processed allele frequency data, typically
to detect genomic regions of interest such as copy number variations or
significant shifts in allele frequency.
}
