% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBSAResults.R
\name{g_statistics_results}
\alias{g_statistics_results}
\title{Calculate g statistics and return a DataFrame of results}
\usage{
g_statistics_results(
  object,
  population_1_sample,
  population_2_sample,
  outlier_filter_method = "delta_snp_index",
  delta_snp_index_threshold = 0.1,
  hampel_filter_threshold = 5.2,
  ...
)
}
\arguments{
\item{object}{A objectxperiment object}

\item{population_1_sample}{One of the samples from object$sample which
will act as the "low bulk" or base sample for delta snp index calculation.}

\item{population_2_sample}{One of the samples from object$sample which
will act as the "high bulk" or comparison sample for delta snp index}
}
\value{
A DataFrame of G statistic results
}
\description{
This is an internal function only -- not exposed in the API. However, it is
accessible indirectly through createBSAResults(). This function calculates
the G statistic for each SNP in the objectxperiment object and returns
a DataFrame of the results.
}
\keyword{internal}
