% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_analysis.R
\name{get_mode}
\alias{get_mode}
\title{Calculate Mode of a Numeric Vector}
\usage{
get_mode(v)
}
\arguments{
\item{v}{Numeric vector.}
}
\value{
The mode of the vector.
}
\description{
Computes the mode of a given numeric vector, which is the value that
appears most frequently.
}
