% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_functions.R
\name{getFDRThreshold}
\alias{getFDRThreshold}
\title{Find false discovery rate threshold}
\usage{
getFDRThreshold(pvalues, alpha = 0.01)
}
\arguments{
\item{pvalues}{}

\item{alpha}{}
}
\value{
The p-value threshold that corresponds to the Benjamini-Hochberg adjusted p-value at the FDR set by alpha.
}
\description{
Find false discovery rate threshold
}
