% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_functions.R
\name{getG}
\alias{getG}
\title{Calculates the G statistic}
\usage{
getG(LowRef, HighRef, LowAlt, HighAlt)
}
\arguments{
\item{LowRef}{A vector of the reference allele depth in the low bulk}

\item{HighRef}{A vector of the reference allele depth in the high bulk}

\item{LowAlt}{A vector of the alternate allele depth in the low bulk}

\item{HighAlt}{A vector of the alternate allele depth in the high bulk}
}
\value{
A vector of G statistic values with the same length as
}
\description{
The function is used by \code{\link{runGprimeAnalysis}} to calculate the G
statisic G is defined by the equation: \deqn{G = 2*\sum_{i=1}^{4}
n_{i}*ln\frac{obs(n_i)}{exp(n_i)}}{G = 2 * \sum n_i * ln(obs(n_i)/exp(n_i))}
Where for each SNP, \eqn{n_i} from i = 1 to 4 corresponds to the reference
and alternate allele depths for each bulk, as described in the following
table: \tabular{rcc}{ Allele \tab High Bulk \tab Low Bulk \cr Reference \tab
\eqn{n_1} \tab \eqn{n_2} \cr Alternate \tab \eqn{n_3} \tab \eqn{n_4} \cr}
...and \eqn{obs(n_i)} are the observed allele depths as described in the data
frame. Method 1 calculates the G statistic using expected values assuming
read depth is equal for all alleles in both bulks: \deqn{exp(n_1) = ((n_1 +
n_2)*(n_1 + n_3))/(n_1 + n_2 + n_3 + n_4)} \deqn{exp(n_2) = ((n_2 + n_1)*(n_2
+ n_4))/(n_1 + n_2 + n_3 + n_4)} etc...
}
\seealso{
\href{https://doi.org/10.1371/journal.pcbi.1002255}{The Statistics
  of Bulk Segregant Analysis Using Next Generation Sequencing}
  \code{\link{tricubeGStat}} for G prime calculation
}
