% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{getG_MH}
\alias{getG_MH}
\title{getG_MH}
\usage{
getG_MH(LowRef, HighRef, LowAlt, HighAlt)
}
\arguments{
\item{LowRef}{LowReference Allele}

\item{HighRef}{HighReference Allele}

\item{LowAlt}{LowAlternate Allele}

\item{HighAlt}{HighAlternate Allele}
}
\description{
Calculates/Returns a G Statistics
}
