% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_functions.R
\name{GetGStat}
\alias{GetGStat}
\title{Calculates the G statistic - method 1}
\usage{
GetGStat(SNPset)
}
\description{
The function is used by \code{\link{ImportFromGATK}} to calculate the G statisic
}
\details{
G is defined by the equation:
\deqn{G = 2*\sum_{i=1}^{4} n_{i}*ln\frac{obs(n_i)}{exp(n_i)}}{G = 2 * \sum n_i * ln(obs(n_i)/exp(n_i))}
Where for each SNP, \eqn{n_i} from i = 1 to 4 corresponds to the reference and
alternate allele depths for each bulk, as described in the following table:
\tabular{rcc}{
Allele \tab High Bulk \tab Low Bulk \cr
Reference \tab \eqn{n_1} \tab \eqn{n_2} \cr
Alternate \tab \eqn{n_3} \tab \eqn{n_4} \cr} ...and \eqn{obs(n_i)} are the observed allele depths as described in the data
frame. In this method for calculating G, the expected values \eqn{exp(n_i)}
are derived by deviding the read depth for the SNP in each bulk by 2. As we
expect 50\% of those reads to support the reference allele.
#' @seealso \code{\link{GetGStat2}}
}
