% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_functions.R
\name{GetGStat2}
\alias{GetGStat2}
\title{Calculates the G statistic - method 2}
\usage{
GetGStat2(SNPset)
}
\arguments{
\item{SNPset}{a data frame with SNPs and genotype fields as imported by ImportFromGATK
G is defined by the equation:
\deqn{G = 2*\sum_{i=1}^{4} n_{i}*ln\frac{obs(n_i)}{exp(n_i)}}{G = 2 * \sum n_i * ln(obs(n_i)/exp(n_i))}
Where for each SNP, \eqn{n_i} from i = 1 to 4 corresponds to the reference and
alternate allele depths for each bulk, as described in the following table:
\tabular{rcc}{
Allele \tab High Bulk \tab Low Bulk \cr
Reference \tab \eqn{n_1} \tab \eqn{n_2} \cr
Alternate \tab \eqn{n_3} \tab \eqn{n_4} \cr} ...and \eqn{obs(n_i)} are the
observed allele depths as described in the data
frame. Method 2 calculates the G statistic using expected values assuming read depth
is equal for all alleles in both bulks: \eqn{((n_1 + n_3)*(n_2 + n_4))/(n_1 + n_2 + n_3 + n_4)}}
}
\description{
The function is used by \code{\link{ImportFromGATK}} to calculate the G statisic
}
\seealso{
\code{\link{GetGStat}}
}
