% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{getPvals_MH}
\alias{getPvals_MH}
\title{getPvals_MH}
\usage{
getPvals_MH(
  Gprime,
  deltaSNP = NULL,
  outlierFilter = c("deltaSNP", "Hampel"),
  filterThreshold
)
}
\arguments{
\item{Gprime}{G Prime Statistic}

\item{deltaSNP}{Default is NULL}

\item{outlierFilter}{Choose outlier filtering method Either deltaSNP or Hampel}

\item{filterThreshold}{Choose filter threshold value should be less than 0.5}
}
\description{
Calculates/Returns pvalues
}
