% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{getQTLTable}
\alias{getQTLTable}
\title{Export a summarized table of QTL}
\usage{
getQTLTable(SNPset, method = "Gprime", alpha = 0.05, interval = 99,
  export = FALSE, fileName = "QTL.csv")
}
\arguments{
\item{SNPset}{Data frame SNP set containing previously filtered SNPs}

\item{method}{either "Gprime" or "QTLseq". The method for detecting significant regions.}

\item{alpha}{numeric. The required false discovery rate alpha for use with \code{method = "Gprime"}}

\item{interval}{numeric. For use eith \code{method = "QTLseq"} The Takagi based confidence interval requested. This will find the column named "CI_\*\*", where \*\* is the requested interval, i.e. 99.}

\item{export}{logical. If TRUE will export a csv table.}

\item{fileName}{either a character string naming a file or a connection open for writing. "" indicates output to the console.}
}
\value{
Returns a summarized table of QTL identified. The table contains the following columns:
\itemize{
\item{id - the QTL identification number}
\item{chromosome - The chromosome on which the region was identified}
\item{start - the start position on that chromosome, i.e. the position of the first SNP that passes the FDR threshold}
\item{end - the end position}
\item{length - the length in basepairs from start to end of the region}
\item{nSNPs - the number of SNPs in the region}
\item{avgSNPs_Mb - the average number of SNPs/Mb within that region}
\item{peakDeltaSNP - the tricube-smoothed deltaSNP-index value at the peak summit}
\item{posPeakDeltaSNP - the position of the absolute maximum tricube-smoothed deltaSNP-index}
\item{maxGprime - the max G' score in the region}
\item{posMaxGprime - the genomic position of the maximum G' value in the QTL}
\item{meanGprime - the average G' score of that region}
\item{sdGprime - the standard deviation of G' within the region}
\item{AUCaT - the Area Under the Curve but above the Threshold line, an indicator of how significant or wide the peak is}
\item{meanPval - the average p-value in the region}
\item{meanQval - the average adjusted p-value in the region}
}
}
\description{
Export a summarized table of QTL
}
