% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hampelFilter.R
\name{hampelFilter}
\alias{hampelFilter}
\title{Temp Docs}
\usage{
hampelFilter(g_prime_vector, hampel_threshold_multiplier = 5.2)
}
\arguments{
\item{threshold_multiplier}{The number of median absolute deviations from the
median to use as a threshold for filtering outlier regions. Default is 5.2}
}
\description{
Temp Docs
}
