% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_hampel_filter.R
\name{hampelFilterGRanges}
\alias{hampelFilterGRanges}
\title{Hampel Filter for Genomic Data in GRanges Object}
\usage{
hampelFilterGRanges(
  gr,
  metric_colname = "metric",
  window_size = 2e+06,
  hampel_threshold_multiplier = 5.2,
  ncpus = 1,
  parallel_backend = NULL
)
}
\arguments{
\item{gr}{A GRanges object containing the genomic data.}

\item{metric_colname}{The name of the metric stored in the metadata columns (mcols) of the GRanges object.}

\item{window_size}{The size of the genomic window used for computing the median and MAD, in base pairs (e.g., 1e6 for 1 Mb).}

\item{hampel_threshold_multiplier}{A numeric value specifying the number of
median absolute deviations from the median to use as a threshold for filtering
outlier regions. Default is 5.2.}
}
\value{
A boolean vector of the same length as the input GRanges object where TRUE
indicates that the corresponding metric value is an outlier.
}
\description{
Applies the Hampel filter to a metric in a GRanges object to identify
outliers. The filter uses genomic windows to compute the median and
the median absolute deviation (MAD) for each sample, then identifies
outliers. See the MATLAB hampel filter documentation for more details:
https://www.mathworks.com/help/dsp/ref/hampelfilter.html
}
\examples{
gr <- GRanges(
  seqnames = c("chr1", "chr1", "chr1", "chr1", "chr1"),
  ranges = IRanges(start = c(1, 2e5, 4e5, 6e5, 8e5), width = 1),
  metric = c(1, 2, 100, 3, 2)
)
outliers <- hampelFilterGRanges(gr, "metric", window_size = 1e6)
print(outliers)

}
