% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Filter.R
\name{ImportFromGATK}
\alias{ImportFromGATK}
\title{Imports SNP data from GATK VarToTable output.}
\usage{
ImportFromGATK(filename, HighBulk = character(), LowBulk = character(),
  ChromList = NULL)
}
\arguments{
\item{filename}{The name of the GATK VarToTable output .table file which the data are to be read from.}

\item{HighBulk}{The sample name of the High Bulk}

\item{LowBulk}{The sample name of the Low Bulk}

\item{ChromList}{a string vector of the chromosomes to be used in the analysis. Useful for filtering out unwanted contigs etc.}
}
\description{
Imports SNP data from GATK VarToTable output.
}
\examples{
df <-  ImportFromGATK(filename = file.table,
    HighBulk = HighBulkSampleName,
    LowBulk = LowBulkSampleName,
    ChromList = c("Chr1","Chr4","Chr7"))
}
