% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_analysis.R
\name{is_noisy}
\alias{is_noisy}
\title{Determine if a Data Point is Noisy}
\usage{
is_noisy(freq_vector, index, neighborhood_size)
}
\arguments{
\item{freq_vector}{Numeric vector of allele frequencies.}

\item{index}{Integer, the index of the data point to check.}

\item{neighborhood_size}{Integer, the number of points on either side of
the index to include in the neighborhood.}
}
\value{
Logical indicating whether the data point is noisy.
}
\description{
Checks whether a data point in a vector of allele frequencies is considered
noisy based on its neighborhood defined by a specified window size.
It evaluates both left and right windows separately and considers the point
noisy only if it does not match the mode of either window.
}
