% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{matrixFilter<-}
\alias{matrixFilter<-}
\alias{matrixFilter}
\alias{matrixFilter<-,BSAExperiment-method}
\alias{matrixFilter,BSAExperiment-method}
\title{Getter and Setter of the BSAExperiment matrixFilter matrix}
\usage{
\S4method{matrixFilter}{BSAExperiment}(object) <- value

\S4method{matrixFilter}{BSAExperiment}(object)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{value}{A boolean matrix of dim .
A value of \code{FALSE} indicates that the corresponding entry of all of the
assay data matrices should be set to NA.}
}
\value{
The setter will return a BSAExperiment object with the matrix
filter matrix added to the assays The getter will return
the boolean matrix
}
\description{
The matrixFilter is a boolean matrix of \verb{nrow(object) x ncol(object)} that
is stored in the assays slot. A value of \code{FALSE} indicates that the
the entry in the other matricies in the assays slot that correspond to
the entry in the matrixFilter should be set to \code{NA}.
}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{applyFilter}()},
\code{\link{comparisons<-}()},
\code{\link{deltaSnpIndex}()},
\code{\link{populationDepths}()},
\code{\link{refDepth}()},
\code{\link{sampleFilter<-}()},
\code{\link{snpIndex}()},
\code{\link{totalDepth<-}()},
\code{\link{variantFilter<-}()}
}
\concept{BSAExperiment-methods}
