% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{Obs_Allele_Freq3}
\alias{Obs_Allele_Freq3}
\title{Obs_Allele_Freq3}
\usage{
Obs_Allele_Freq3(SNPSet)
}
\arguments{
\item{SNPSet}{A SNPSet generated from the function importFromTable}
}
\description{
Returns a 4 column data filtering for specific Chromosome and High Bulk Observed Allele Frequencies
}
