% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination_analysis.R
\name{plot_breakpoints}
\alias{plot_breakpoints}
\title{Plot Breakpoints on Allele Frequency Data}
\usage{
plot_breakpoints(
  allele_freq_df,
  breakpoint_df,
  allele_freq_stat,
  de_noise = FALSE,
  sample_set = c(),
  seqnames_set = c(),
  point_size = 0.7
)
}
\arguments{
\item{allele_freq_df}{A data frame containing allele frequency data.}

\item{breakpoint_df}{A data frame containing breakpoint data.}

\item{allele_freq_stat}{Character string specifying the column name of the allele frequency
statistic to plot. This should be an existing column in \code{allele_freq_df}.}

\item{sample_set}{Character vector specifying the samples to include in the plot.
If empty, all samples will be included.}

\item{seqnames_set}{Character vector specifying the seqnames to include in the plot.
If empty, all seqnames will be included.}
}
\value{
A ggplot object displaying the allele frequency data with
breakpoints overlaid, facilitating visual assessment of genomic variations.
}
\description{
This function creates a plot of allele frequency data for each sample and chromosome,
with breakpoints overlaid as dashed vertical lines.
}
