% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotGprimeDist}
\alias{plotGprimeDist}
\title{Plots Gprime distribution}
\usage{
plotGprimeDist(SNPset, outlierFilter = c("deltaSNP", "Hampel"))
}
\arguments{
\item{SNPset}{a data frame with SNPs and genotype fields as imported by
\code{ImportFromGATK} and after running \code{GetPrimeStats}}
}
\value{
Plots a ggplot histogram of the G' value distribution. It will then
overlay an estimated log normal distribution with the same mean and variance
as your G' distribution. This will allow to verify if after filtering your G'
value appear to be close to log normally and thus can be used to estimate
p-values using the non-parametric estimation method described in Magwene et al. (2013). Breifly,
using the natural log of Gprime a median absolute deviation (MAD) is
calculated. The Gprime set is trimmed to exclude outlier regions (i.e. QTL)
based on Hampel's rule. An estimation of the mode of the trimmed set is
calculated using the \code{\link[modeest]{mlv}} function from the package modeest. Finally, the mean
and variance of the set are estimated using the median and mode are
estimated and used to plot the log normal distribution.
}
\description{
Plots a ggplot histogram of the distribution of Gprime with a log normal
distribution overlay
}
\examples{
plotGprimedist(df_filt_6Mb, outlierFilter = "deltaSNP")

}
\seealso{
\code{\link{GetPvals}} for how p-values are calculated.
}
