\name{plotGprimeDist_Py_MH}
\alias{plotGprimeDist_Py_MH}
\title{plotGprimeDist_Py_MH}
\usage{
  plotGprimeDist_Py_MH(SNPset, outlierFilter = c("deltaSNP", "Hampel"), filterThreshold = 0.1, binwidth = 0.5)
}
\arguments{
  \item{SNPset}{An SNPset generated from importFromTable function}
  
  \item{outlierFilter}{AN outlier filter method either deltaSNP or Hampel}
  
  \item{filterThreshold}{Choose filterThreshold Value}
  
  \item{binwidth}{Choose a binwidth size}
  
}
\description{
  Calculates G Prime Distribution in jupyter notebook for you
}
