% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{plotQTLStats_MH2}
\alias{plotQTLStats_MH2}
\title{Plots different paramaters for QTL identification}
\usage{
plotQTLStats_MH2(
  SNPset,
  subset = NULL,
  var = "diff",
  scaleChroms = TRUE,
  line = TRUE,
  ...
)
}
\arguments{
\item{SNPset}{a data frame with SNPs and genotype fields as imported by
\code{ImportFromGATK} and after running \code{runGprimeAnalysis} or \code{runQTLseqAnalysis}}

\item{subset}{a vector of chromosome names for use in quick plotting of
chromosomes of interest. Defaults to
NULL and will plot all chromosomes in the SNPset}

\item{var}{character. "diff"}

\item{scaleChroms}{boolean. if TRUE (default) then chromosome facets will be
scaled to relative chromosome sizes. If FALSE all facets will be equal
sizes. This is basically a convenience argument for setting both scales and
shape as "free_x" in ggplot2::facet_grid.}

\item{line}{boolean. If TRUE will plot line graph. If FALSE will plot points.
Plotting points will take more time.}

\item{...}{arguments to pass to ggplot2::geom_line or ggplot2::geom_point for
changing colors etc.}
}
\value{
Plots a ggplot graph for all chromosomes or those requested in
  \code{subset}. By setting \code{var} to "nSNPs" the distribution of SNPs
  used to calculate G' will be plotted. "deltaSNP" will plot a tri-cube
  weighted delta SNP-index for each SNP. "Gprime" will plot the tri-cube
  weighted G' value. Setting "negLogPval" will plot the -log10 of the p-value
  at each SNP. In "Gprime" and "negLogPval" plots, a genome wide FDR threshold of
  q can be drawn by setting "plotThreshold" to TRUE. The defualt is a red
  line. If you would like to plot a different line we suggest setting
  "plotThreshold" to FALSE and manually adding a line using
  ggplot2::geom_hline.
}
\description{
A wrapper for ggplot to plot genome wide distribution of parameters used to
identify QTL.
}
