% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotSimulatedThresholds}
\alias{plotSimulatedThresholds}
\title{Plots simulation data for QTLseq analysis}
\usage{
plotSimulatedThresholds(SNPset = NULL, popStruc = "F2", bulkSize,
  depth = NULL, replications = 10000, filter = 0.3, intervals = c(95,
  99))
}
\arguments{
\item{SNPset}{optional. Either supply your data set to extract read depths from or supply depth vector.}

\item{popStruc}{the population structure. Defaults to "F2" and assumes "RIL" otherwise.}

\item{bulkSize}{non-negative integer. The number of individuals in each bulk}

\item{depth}{optional integer vector. A read depth for which to replicate SNP-index calls. If read depth is defined SNPset will be ignored.}

\item{replications}{integer. The number of bootstrap replications.}

\item{filter}{numeric. An optional minimum SNP-index filter}

\item{intervals}{numeric vector. Confidence intervals supplied as two-sided percentiles. i.e. If intervals = '95' will return the two sided 95\% confidence interval, 2.5\% on each side.}
}
\value{
Plots a deltaSNP by depth plot. Helps if the user wants to know the the delta SNP index needed to pass a certain CI at a specified depth.
}
\description{
as described in Takagi et al., (2013). Genotypes are randomly assigned for
each indvidual in the bulk, based on the population structure. The total
alternative allele frequency in each bulk is calculated at each depth used to simulate
delta SNP-indeces, with a user defined number of bootstrapped replication.
The requested confidence intervals are then calculated from the bootstraps.
This function plots the simulated confidence intervals by the read depth.
}
\examples{
plotSimulatedThresholds <- function(SNPset = NULL, popStruc = "F2", bulkSize = 25,   depth = 1:150, replications = 10000, filter = 0.3, intervals = c(95, 99))
}
