% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{populationDepths}
\alias{populationDepths}
\alias{populationDepths,BSAExperiment-method}
\title{Extract population alt and ref allele depths from a BSAExperiment object}
\usage{
\S4method{populationDepths}{BSAExperiment}(object, population_1_sample, population_2_sample)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{population_1_sample}{The name of the sample used as the first
population. In the context of the Takagi 2013 paper, this would be the
'low bulk' sample.}

\item{population_2_sample}{The name of the sample used for the second
population. In the context of the Takagi 2013 paper, this would be the
'high bulk' sample.}
}
\description{
Extract population alt and ref allele depths from a BSAExperiment object
}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{applyFilter}()},
\code{\link{comparisons<-}()},
\code{\link{deltaSnpIndex}()},
\code{\link{matrixFilter<-}()},
\code{\link{refDepth}()},
\code{\link{sampleFilter<-}()},
\code{\link{snpIndex}()},
\code{\link{totalDepth<-}()},
\code{\link{variantFilter<-}()}
}
\concept{BSAExperiment-methods}
