% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{QTLParser_1_MH}
\alias{QTLParser_1_MH}
\title{QTLParser_1_MH}
\usage{
QTLParser_1_MH(vcf, HighBulk, LowBulk, filename)
}
\arguments{
\item{vcf}{A vcf file that is as a tidy data frame}

\item{HighBulk}{The name of the HighBulk sample}

\item{LowBulk}{The name of the LowBulk sample}

\item{filename}{Provide a partiular name to identify input CSV file}
}
\description{
Parses a vcf file which has been converted to a tidy data frame before running the operation
}
