\name{QTLParser_2_MH}
\alias{QTLParser_2_MH}
\title{QTLParser_2_MH}
\usage{
  QTLParser_1_MH(vcf, HighBulk, LowBulk, RootDirectory)
}
\arguments{
  \item{vcf}{A vcf file that is as a tidy data frame}
  
  \item{HighBulk}{The name of the HighBulk sample}
  
  \item{LowBulk}{The name of the LowBulk sample}
  
  \item{RootDirectory}{The name of the root directory where you cloned this repository from Github}
}
\description{
  Parses a vcf file which has been converted to a tidy data frame before running the operation
}
