% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{QTLseqr-package}
\alias{QTLseqr}
\alias{QTLseqr-package}
\title{QTLseqr: QTL mapping using Bulk Segregant Analysis of Next Generation Sequencing data.}
\description{
QTLseqr performs QTL mapping using Bulk Segregant Analysis of Next Gen Sequencing data. The package is an R implementation of the analyses described in Magwene PM, Willis JH, Kelly JK (2011) The Statistics of Bulk Segregant Analysis Using Next Generation Sequencing. PLOS Computational Biology 7(11): e1002255. doi: 10.1371/journal.pcbi.1002255 and Takagi, H., Abe, A., Yoshida, K., Kosugi, S., Natsume, S., Mitsuoka, C., Uemura, A., Utsushi,H., Tamiru, M., Takuno, S., Innan, H., Cano, L. M., Kamoun, S. and Terauchi, R. (2013), QTL-seq: rapid mapping of quantitative trait loci in rice by whole genome resequencing of DNA from two bulked populations. *Plant J*, 74: 174–183. doi:10.1111/tpj.12105
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bmansfeld/QTLseqr/}
  \item Report bugs at \url{https://github.com/bmansfeld/QTLseqr/issues}
}

}
\author{
\strong{Maintainer}: Ben Nathan Mansfeld \email{mansfeld@msu.edu}

Authors:
\itemize{
  \item Chase Mateusiak \email{chasem@wustl.edu} (\href{https://orcid.org/0000-0002-2890-4242}{ORCID})
}

Other contributors:
\itemize{
  \item Rebecca Grumet \email{grumet@msu.edu} [thesis advisor]
}

}
\keyword{internal}
