% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{refDepth}
\alias{refDepth}
\alias{refDepth,BSAExperiment-method}
\title{Get the Reference Depths from the BSAExperiment}
\usage{
\S4method{refDepth}{BSAExperiment}(object, sample_set = NULL)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{sample_set}{A character vector of samples, possibly of length 1.
All entries in \code{sample_set} must be present in the \code{sample} column of
the colData}
}
\value{
A matrix of total depths
}
\description{
Optionally, a vector of BSAEXperiment colnames, possibly of length 1,
may be passed to extract a subset of the samples from the BSAExperiment
object
}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{applyFilter}()},
\code{\link{comparisons<-}()},
\code{\link{deltaSnpIndex}()},
\code{\link{matrixFilter<-}()},
\code{\link{populationDepths}()},
\code{\link{sampleFilter<-}()},
\code{\link{snpIndex}()},
\code{\link{totalDepth<-}()},
\code{\link{variantFilter<-}()}
}
\concept{BSAExperiment-methods}
