% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{runGprimeAnalysis_GPrime_Smooth}
\alias{runGprimeAnalysis_GPrime_Smooth}
\title{runGprimeAnalysis_GPrime_Smooth}
\usage{
runGprimeAnalysis_GPrime_Smooth(
  SNPset,
  windowSize = windowSize,
  outlierFilter = "deltaSNP",
  filterThreshold = 0.1,
  deg = deg,
  nn = nn
)
}
\arguments{
\item{SNPset}{An SNPset}

\item{windowSize}{Specify the WindowSize}

\item{outlierFilter}{Specify the outlierFilter}

\item{filterThreshold}{Specify threshold filter value default is 0.1}

\item{deg}{Degree of Polynomial to use for LocFit Model}

\item{nn}{Nearest Neighbor component of smoothing parameter. Default value is 0.7, unless h or adpen are provided, in which case the default is 0}
}
\description{
Runs a Gprime Analysis on Bulk segregants with additional smoothing parameters for G Prime Test Statistic
}
