% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{runGprimeAnalysis_MH}
\alias{runGprimeAnalysis_MH}
\title{runGprimeAnalysis_MH}
\usage{
runGprimeAnalysis_MH(
  SNPset,
  windowSize = 1e+06,
  outlierFilter = "deltaSNP",
  filterThreshold = 0.1,
  ...
)
}
\arguments{
\item{SNPset}{An SNPset}

\item{windowSize}{Specify the WindowSize}

\item{outlierFilter}{Specify the outlierFilter}

\item{filterThreshold}{Specify threshold filter value default is 0.1}
}
\description{
Runs a Gprime Analysis on Bulk segregants
}
