% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takagi_sim.R
\name{runQTLseqAnalysis}
\alias{runQTLseqAnalysis}
\title{Calculates delta SNP confidence intervals for QTLseq analysis}
\usage{
runQTLseqAnalysis(SNPset, windowSize = 1e+06, popStruc = "F2", bulkSize,
  depth = NULL, replications = 10000, filter = 0.3, intervals = c(95,
  99))
}
\arguments{
\item{SNPset}{The data frame imported by \code{ImportFromGATK}}

\item{popStruc}{the population structure. Defaults to "F2" and assumes "RIL"}

\item{bulkSize}{non-negative integer. The number of individuals in each bulk}

\item{depth}{integer. A read depth for which to replicate SNP-index calls.}

\item{replications}{integer. The number of bootstrap replications.}

\item{filter}{numeric. An optional minimum SNP-index filter}

\item{intervals}{numeric vector of probabilities with values in [0,1] 
corresponding to the requested confidence intervals}
}
\value{
A SNPset data frame with delta SNP-index thresholds corrisponding to the 
requested confidence intervals matching the tricube smoothed depth at each SNP.
}
\description{
The method for simulating delta SNP-index confidence interval thresholds
as described in Takagi et al., (2013). Genotypes are randomly assigned for
each indvidual in the bulk, based on the population structure. The total
alternative allele frequency in each bulk is calculated at each depth used to simulate 
delta SNP-indeces, with a user defined number of bootstrapped replication.
The requested confidence intervals are then calculated from the bootstraps.
}
\examples{
df_filt <- runQTLseqAnalysis(
SNPset = df_filt,
windowSize = 1e6,
popStruc = "F2",
replications = 10000,
intervals = c(95, 99)
)


}
