% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{runQTLseqAnalysis_MH}
\alias{runQTLseqAnalysis_MH}
\title{runQTLseqAnalysis_MH}
\usage{
runQTLseqAnalysis_MH(
  SNPset,
  windowSize = 1e+06,
  popStruc = "F2",
  bulkSize,
  depth = NULL,
  replications = 10000,
  filter = 0.3,
  intervals = c(95, 99),
  ...
)
}
\arguments{
\item{SNPset}{A SNPset}

\item{windowSize}{Specify Window Size}

\item{popStruc}{population Structure of the Segregant Bulks "F2" is default}

\item{bulkSize}{Choose appropriate parameters for bulk size in each BSA group}

\item{depth}{Default is NULL}

\item{replications}{Choose the number of replications default is 10000}

\item{filter}{Choose specific filtering value default is 0.3}

\item{intervals}{Choose appropriate confidence intervals default is 95% - 99%}
}
\description{
Calculates Returns a SNPset
}
