% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{sampleFilter<-}
\alias{sampleFilter<-}
\alias{sampleFilter}
\alias{sampleFilter<-,BSAExperiment-method}
\alias{sampleFilter,BSAExperiment-method}
\title{Getter and Setter of the BSAExperiment sampleFilter vector}
\usage{
\S4method{sampleFilter}{BSAExperiment}(object) <- value

\S4method{sampleFilter}{BSAExperiment}(object)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{value}{A boolean vector of length equal to the columns of the
BSAExperiment object. A value of \code{FALSE} indicates that that sample
should be removed from the filtered dataset.}
}
\value{
The setter will return a BSAExperiment object with the sample
filter vector added to the colData. The getter will return
the boolean vector
}
\description{
The sampleFilter is a boolean vector which is added as a column to the
colData. A value of \code{FALSE} indicates that that entire sample should
be removed from the filtered dataset. When passing a vector to the setter,
it must be equal to the number of columns in the BSAExperiment object
}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{applyFilter}()},
\code{\link{comparisons<-}()},
\code{\link{deltaSnpIndex}()},
\code{\link{matrixFilter<-}()},
\code{\link{populationDepths}()},
\code{\link{refDepth}()},
\code{\link{snpIndex}()},
\code{\link{totalDepth<-}()},
\code{\link{variantFilter<-}()}
}
\concept{BSAExperiment-methods}
