% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takagi_sim.R
\name{simulateAlleleFreq}
\alias{simulateAlleleFreq}
\title{Randomly calculates an alternate allele frequency within a bulk}
\usage{
simulateAlleleFreq(n, pop = "F2")
}
\arguments{
\item{n}{non-negative integer. The number of individuals in each bulk}

\item{pop}{the population structure. Defaults to "F2" and assumes "RIL" 
population otherwise.}
}
\value{
an alternate allele frequency withing the bulk. Used for simulating 
SNP-indeces.
}
\description{
Randomly calculates an alternate allele frequency within a bulk
}
