% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takagi_sim.R
\name{simulateConfInt}
\alias{simulateConfInt}
\title{Simulation of delta SPP index confidence intervals}
\usage{
simulateConfInt(popStruc = "F2", bulkSize, depth = 1:100,
  replications = 10000, filter = 0.3, intervals = c(0.005, 0.025, 0.05,
  0.95, 0.975, 0.995))
}
\arguments{
\item{popStruc}{the population structure. Defaults to "F2" and assumes "RIL"}

\item{bulkSize}{non-negative integer. The number of individuals in each bulk}

\item{depth}{integer. A read depth for which to replicate SNP-index calls.}

\item{replications}{integer. The number of bootstrap replications.}

\item{filter}{numeric. An optional minimum SNP-index filter}

\item{intervals}{numeric vector of probabilities with values in [0,1] 
corresponding to the requested confidence intervals}
}
\value{
A data frame of delta SNP-index thresholds corrisponding to the 
requested confidence intervals at the user set depths.
}
\description{
The method for simulating delta SNP-index confidence interval thresholds
as described in Takagi et al., (2013). Genotypes are randomly assigned for
each indvidual in the bulk, based on the population structure. The total
alternative allele frequency in each bulk is calculated at each depth used to simulate 
delta SNP-indeces, with a user defined number of bootstrapped replication.
The requested confidence intervals are then calculated from the bootstraps.
}
\examples{
CI <-
simulateConfInt(
   popStruc = "F2",
   bulkSize = 50,
   depth = 1:100,
   intervals = c(0.05, 0.95, 0.025, 0.975, 0.005, 0.995, 0.0025, 0.9975)
}
