% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takagi_simulation.R
\name{simulateDeltaSnpIndex}
\alias{simulateDeltaSnpIndex}
\title{Simulates the change in Alternate Allele Frequency between population_2
and population_1}
\usage{
simulateDeltaSnpIndex(
  depth,
  population_1_snp_index,
  population_2_snp_index,
  replications = 10000,
  snp_index_filter = NULL
)
}
\arguments{
\item{depth}{An integer vector specifying the read depths at which to
simulate alternate allele frequency calls. Defaults to a sequence from
1 to 100.}

\item{population_1_snp_index}{numeric. The alternate allele frequency for
population 1.}

\item{population_2_snp_index}{numeric. The alternate allele frequency for
population 2.}

\item{replications}{An integer specifying the number of bootstrap
replications to perform. This affects the robustness and accuracy of the
simulated confidence intervals. Defaults to 10000.}

\item{snp_index_filter}{A numeric value specifying the minimum alt
allele frequency filter threshold. Used to exclude low alt allele
frequency values from the simulation. Defaults to 0.3.}
}
\value{
Returns a vector of length replicates of simulated
\eqn{\Delta}{Delta}(Alt Allele Frequency)
}
\description{
The function simulates the change in alternate allele frequency between two
populations with replication
}
