% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takagi_simulation.R
\name{simulateDeltaSnpIndexCI}
\alias{simulateDeltaSnpIndexCI}
\title{Simulate \eqn{\Delta}{Delta}(Alt Allele Frequency) Confidence Intervals}
\usage{
simulateDeltaSnpIndexCI(
  population_1_n,
  population_structure,
  population_2_n = population_1_n,
  depth_vector = 1:100,
  replications = 10000,
  snp_index_filter = 0.3,
  ci_lower_bounds = c(0.05, 0.025)
)
}
\arguments{
\item{population_1_n}{A positive integer specifying the number of
individuals in the first bulk.}

\item{population_structure}{A string indicating the population structure of
the individuals. Valid options are 'F2' for an F2 population and 'RIL' for
Recombinant Inbred Lines. This parameter influences how alternate allele
frequencies are simulated.}

\item{population_2_n}{A positive integer specifying the number of
individuals in the second bulk. Defaults to the same number as
\code{population_1_n}.}

\item{replications}{An integer specifying the number of bootstrap
replications to perform. This affects the robustness and accuracy of the
simulated confidence intervals. Defaults to 10000.}

\item{snp_index_filter}{A numeric value specifying the minimum alt
allele frequency filter threshold. Used to exclude low alt allele
frequency values from the simulation. Defaults to 0.3.}

\item{ci_lower_bounds}{A numeric vector of probabilities with values in \link{0,0.5},
specifying the confidence levels for which to calculate confidence
intervals. For example, c(0.05, 0.025) corresponds to the 95\% and 97.5\%
confidence intervals. Defaults to c(0.05, 0.025).}

\item{depth}{An integer vector specifying the read depths at which to
simulate alternate allele frequency calls. Defaults to a sequence from
1 to 100.}
}
\value{
A data frame with each row representing a specified depth and each
column representing a \eqn{\Delta}{Delta}(Alt Allele Frequency) threshold for the
corresponding confidence interval at that depth. The first column
specifies the depth, and subsequent columns are named according to the
specified confidence intervals.
}
\description{
This function simulates \eqn{\Delta}{Delta}(Alt Allele Frequency) confidence
intervals for two populations of individuals, given a specified population
structure. It calculates alternate allele frequencies within each bulk
and then simulates alternate allele frequency values across a range of
depths to determine confidence intervals for the
\eqn{\Delta}{Delta}(Alt Allele Frequency). This process is repeated for
a specified number of replications to ensure robustness of the confidence
interval estimates.
}
\examples{
simulateDeltaSnpIndexCI(
  population_1_n = 50,
  population_structure = "F2",
  depth = 1:10,
  replications = 100,
  filter = 0.3,
  intervals = c(0.05, 0.025)
)
}
