% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePopulationAltFrequency.R
\name{simulatePopulationAltFrequency}
\alias{simulatePopulationAltFrequency}
\title{Simulate a Population's Alternate Allele Frequency}
\usage{
simulatePopulationAltFrequency(n, population_structure)
}
\arguments{
\item{n}{non-negative integer. The number of individuals in the population.}

\item{population_structure}{A string indicating the population structure of
the individuals. Valid options are 'F2' for an F2 population and 'RIL' for
Recombinant Inbred Lines. This parameter influences how alternate allele
frequencies are simulated.}
}
\value{
Returns a numeric value representing the calculated alternate allele
frequency within the population. This value is a proportion ranging from
0 to 1, where 0 indicates no presence of the alternate allele and 1
indicates complete presence.
}
\description{
This function calculates the alternate allele frequency for a given number of
individuals in a population, based on the specified population structure.
It supports two population structures: F2 and RIL. For an F2 population,
alleles are sampled with probabilities reflecting Mendelian inheritance
for a single heterozygous locus. For RILs (Recombinant Inbred Lines),
alleles are sampled assuming an equal probability of inheritance from
either parent.
}
