% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takagi_sim.R
\name{simulateSNPindex}
\alias{simulateSNPindex}
\title{Simulates a delta SNP-index with replication}
\usage{
simulateSNPindex(depth, altFreq1, altFreq2, replicates = 10000,
  filter = NULL)
}
\arguments{
\item{depth}{integer. A read depth for which to replicate SNP-index calls.}

\item{altFreq1}{numeric. The alternate allele frequency for bulk A.}

\item{altFreq2}{numeric. The alternate allele frequency for bulk B.}

\item{replicates}{integer. The number of bootstrap replications.}

\item{filter}{numeric. an optional minimum SNP-index filter}
}
\value{
Returns a vector of length replicates delta SNP-indeces
}
\description{
Simulates a delta SNP-index with replication
}
