% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothAlleleFrequencyMetric.R
\name{smoothAlleleFrequencyMetric}
\alias{smoothAlleleFrequencyMetric}
\title{Calculate tricube weighted statistics for each allele}
\usage{
smoothAlleleFrequencyMetric(allele_row_ranges, stat, window_size = 2e+06, ...)
}
\arguments{
\item{allele_row_ranges}{A GRangesList object where each record represents
The genomic coordinate of a allele}

\item{stat}{A vector of statistics calculated from the alleles. This might be
the G statistic, see \code{\link{calculateGStatistic}} or
the snp_index from \code{\link{populationDepths}}, for example.
The order of this vector must correspond with the order of the
allele_row_ranges object.}

\item{window_size}{the window size (in base pairs) bracketing each allele for which
to calculate the statitics. Magwene et. al recommend a window size of ~25
cM, but also recommend optionally trying several window sizes to test if
peaks are over- or undersmoothed.}

\item{...}{Other arguments passed to \code{\link[locfit]{locfit}} and
subsequently to \code{\link[locfit]{locfit.raw}}() (or the lfproc). Usefull
in cases where you get "out of vertex space warnings"; Set the maxk higher
than the default 100. See \code{\link[locfit]{locfit.raw}}().}
}
\value{
Returns a vector of the weighted statistic caluculted with a tricube
smoothing kernel
}
\description{
Uses local regression (wrapper for \code{\link[locfit]{locfit}}) to predict a
tricube smoothed version of the statistic supplied for each allele. This works as a
weighted average across neighboring alleles that accounts for Linkage
disequilibrium (LD) while minizing noise attributed to allele calling errors.
Values for neighboring alleles within the window are weighted by physical
distance from the focal allele.
}
\seealso{
\code{\link{getG}} for G statistic calculation

\code{\link[locfit]{locfit}} for local regression
}
