% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{snpIndex}
\alias{snpIndex}
\alias{snpIndex,BSAExperiment-method}
\title{Calculate the SNP index of a given sample}
\usage{
\S4method{snpIndex}{BSAExperiment}(object, sample_set = NULL)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{sample_set}{A character vector of samples, possibly of length 1.
All entries in \code{sample_set} must be present in the \code{sample} column of
the colData}
}
\description{
Calculate the SNP index of a given sample
}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{applyFilter}()},
\code{\link{comparisons<-}()},
\code{\link{deltaSnpIndex}()},
\code{\link{matrixFilter<-}()},
\code{\link{populationDepths}()},
\code{\link{refDepth}()},
\code{\link{sampleFilter<-}()},
\code{\link{totalDepth<-}()},
\code{\link{variantFilter<-}()}
}
\concept{BSAExperiment-methods}
