% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary,BSAResults-method}
\alias{summary,BSAResults-method}
\title{Summarize a BSAResults object}
\usage{
\S4method{summary}{BSAResults}(object)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}
}
\description{
Summarize a BSAResults object
}
