% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{totalDepth<-}
\alias{totalDepth<-}
\alias{totalDepth}
\alias{totalDepth<-,BSAExperiment-method}
\alias{totalDepth,BSAExperiment-method}
\title{Getter/Setter of the BSAExperiment totalDepth assay matrix}
\usage{
\S4method{totalDepth}{BSAExperiment}(object) <- value

\S4method{totalDepth}{BSAExperiment}(object, sample_set = NULL)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{value}{A numeric matrix storing the total depth at each variant
location in the BSAExperiment. The dimensions of this matrix must be
equal to the dimension of the BSAExperiment, and must have the same
row and column names}

\item{sample_set}{A character vector of samples, possibly of length 1.
All entries in \code{sample_set} must be present in the \code{sample} column of
the colData}
}
\value{
The setter will return a BSAExperiment object with the
total depth slot updated. The getter will return a matrix of
total depths
}
\description{
Set the totalDepth matrix in the BSAExperiment object, or retrieve it.
For the getter, a vector of BSAEXperiment colnames, possibly of length 1,
may be passed to extract a subset of the samples from the BSAExperiment
object
}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{applyFilter}()},
\code{\link{comparisons<-}()},
\code{\link{deltaSnpIndex}()},
\code{\link{matrixFilter<-}()},
\code{\link{populationDepths}()},
\code{\link{refDepth}()},
\code{\link{sampleFilter<-}()},
\code{\link{snpIndex}()},
\code{\link{variantFilter<-}()}
}
\concept{BSAExperiment-methods}
