% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{tricube_Smooth}
\alias{tricube_Smooth}
\title{tricube_Smooth}
\usage{
tricube_Smooth(POS, Stat, windowSize, deg, nn)
}
\arguments{
\item{POS}{THe Position vector of where the SNPs were called}

\item{Stat}{The G Statistic, however, it is the G~lp(x) for a local regression and likelihood model}

\item{windowSize}{Specify the WindowSize}

\item{deg}{Degree of Polynomial to use for LocFit Model}

\item{nn}{Nearest Neighbor component of smoothing parameter. Default value is 0.7, unless h or adpen are provided, in which case the default is 0}
}
\description{
Applies Tricube Kernel to G Test Statistic
}
