% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_functions.R
\name{tricubeStat}
\alias{tricubeStat}
\title{Calculate tricube weighted statistics for each SNP}
\usage{
tricubeStat(POS, Stat, windowSize = 2e+06)
}
\arguments{
\item{POS}{A vector of genomic positions for each SNP}

\item{Stat}{A vector of values for a given statistic for each SNP}

\item{WinSize}{the window size (in base pairs) bracketing each SNP for which 
to calculate the statitics. Magwene et. al recommend a window size of ~25 
cM, but also recommend optionally trying several window sizes to test if 
peaks are over- or undersmoothed.}
}
\value{
Returns a vector of the weighted statistic caluculted with a
  tricube smoothing kernel
}
\description{
Uses local regression (wrapper for \code{\link[locfit]{locfit}}) to predict a 
tricube smoothed version of the statistic supplied for each SNP. This works as a
weighted average across neighboring SNPs that accounts for Linkage
disequilibrium (LD) while minizing noise attributed to SNP calling errors.
Values for neighboring SNPs within the window are weighted by physical
distance from the focal SNP.
}
\examples{
df_filt_4mb$Gprime <- tricubeStat(POS, Stat = GStat, WinSize = 4e6)
}
\seealso{
\code{\link{getG}} for G statistic calculation

\code{\link[locfit]{locfit}} for local regression
}
