% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{tricubeStat_MH}
\alias{tricubeStat_MH}
\title{tricubeStat_MH}
\usage{
tricubeStat_MH(POS, Stat, windowSize = 2e+06, ...)
}
\arguments{
\item{POS}{An SNPset}

\item{Stat}{Specify the WindowSize}

\item{windowSize}{Specify the outlierFilter}
}
\description{
Delivers a tricubeStat function
}
