% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.R
\name{validatePositiveInteger}
\alias{validatePositiveInteger}
\title{Validate that a given variable is a positive integer}
\usage{
validatePositiveInteger(n, varname = "Variable")
}
\arguments{
\item{n}{The variable to validate}

\item{varname}{The name of the variable to include in the error message}
}
\description{
Validate that a given variable is a positive integer
}
\keyword{internal}
