% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{variantFilter<-}
\alias{variantFilter<-}
\alias{variantFilter}
\alias{variantFilter<-,BSAExperiment-method}
\alias{variantFilter,BSAExperiment-method}
\title{Getter/Setter of the BSAExperiment variantFilter vector}
\usage{
\S4method{variantFilter}{BSAExperiment}(object) <- value

\S4method{variantFilter}{BSAExperiment}(object)
}
\arguments{
\item{object}{A \code{BSAExperiment} object}

\item{value}{A boolean vector of length equal to the rows of the
BSAExperiment object. A value of \code{FALSE} indicates that that variant
should be removed from the filtered dataset.}
}
\value{
The setter will return a BSAExperiment object with the variant
filter vector added to the mcols of the rowRanges. The getter will return
the boolean vector
}
\description{
The variantFilter is a boolean vector which is added as a column to the
rowRanges mcols. A value of \code{FALSE} indicates that that variant should
be removed from the filtered dataset. When passing a vector to the setter,
it must be equal to the number of rows in the BSAExperiment object
}
\seealso{
Other BSAExperiment-methods: 
\code{\link{altDepth<-}()},
\code{\link{applyFilter}()},
\code{\link{comparisons<-}()},
\code{\link{deltaSnpIndex}()},
\code{\link{matrixFilter<-}()},
\code{\link{populationDepths}()},
\code{\link{refDepth}()},
\code{\link{sampleFilter<-}()},
\code{\link{snpIndex}()},
\code{\link{totalDepth<-}()}
}
\concept{BSAExperiment-methods}
