% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{AlleleFreqSlidingWindow}
\alias{AlleleFreqSlidingWindow}
\title{AlleleFreqSlidingWindow}
\usage{
AlleleFreqSlidingWindow(
  vcf,
  chromList = NULL,
  windowSize = NULL,
  highBulk = NULL,
  lowBulk = NULL,
  filename = NULL
)
}
\arguments{
\item{vcf}{A vcf file}

\item{windowSize}{Specify window size to calculate number of SNPs and Tricube Stat Window Size}

\item{highBulk}{Specify sample name from vcf of the high Bulk}

\item{lowBulk}{Specify sample name from vcf of the lowBulk}

\item{filename}{Specify the file name used in nested importFromVCF function}

\item{chromlist}{A vector specifying particular chromosomes}
}
\description{
AlleleFreqSlidingWindow
}
