\name{Correlation}
\alias{Correlation}
\title{Correlation}
\usage{
  Correlation(file, chromlist = NULL,p1 = NULL, p2 = NULL, p3 = NULL, p4 = NULL,p5=TRUE) 
}
\arguments{
  \item{file}{A vcf file}
  \item{chromlist}{A vector of Chromosome/Contig names reverse compatible with VCF file}
  \item{p1}{Boolean TRUE or FALSE}
  \item{p2}{Boolean TRUE or FALSE}
  \item{p3}{Boolean TRUE or FALSE)
  \item{p4}{Boolean TRUE or FALSE}
  \item{p5}{Boolean TRUE or FALSE}
  }
  
}
\description{
  Returns correlation values, p-values, a heatmap, and correlogram 
}
