% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBSAResults.R
\name{createBSAResults}
\alias{createBSAResults}
\title{Briefly, using the natural log of Gprime a median absolute deviation (MAD)
is calculated. The Gprime set is trimmed to exclude outlier regions
(i.e. QTL) based on Hampel's rule. An alternate method for filtering out
QTL is proposed using absolute delta SNP indeces greater than
a set threshold to filter out potential QTL. An estimation of the mode of the trimmed set}
\usage{
createBSAResults(
  comparison_index,
  bsae,
  filter_method,
  window_size = 1e+06,
  delta_alt_frequency_filter = 0.1,
  hampel_threshold_multiplier = 5.2,
  ...
)
}
\description{
Briefly, using the natural log of Gprime a median absolute deviation (MAD)
is calculated. The Gprime set is trimmed to exclude outlier regions
(i.e. QTL) based on Hampel's rule. An alternate method for filtering out
QTL is proposed using absolute delta SNP indeces greater than
a set threshold to filter out potential QTL. An estimation of the mode of the trimmed set
}
