% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{Facet_Allelic_Chrom}
\alias{Facet_Allelic_Chrom}
\title{Facet_Allelic_Chrom}
\usage{
Facet_Allelic_Chrom(
  SNPset,
  subset = NULL,
  var1 = "AD_ALT.LOW",
  var2 = "AD_ALT.HIGH",
  scaleChroms = TRUE,
  line = TRUE
)
}
\arguments{
\item{SNPset}{A vcf file that has been imported by ImportVCF, filtered by SNPfilter, and run}

\item{subset}{A Chromosome subset}

\item{var1}{This is a variable that you cannot change, it must be AD_ALT.LOW}

\item{var2}{This is a variable that you cannot change, it must be AD_ALT.HIGH}

\item{scaleChroms}{TRUE or FASLE do you want plots on each Chromosome or NOT}

\item{line}{TRUE or FALSE do you want a line graph or just points}
}
\description{
Provides plot of Alterante allelic depths for High and Low Bulk, sorry no sliding window.
}
