% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Filter.R
\name{FilterSNPs}
\alias{FilterSNPs}
\title{Filter SNPs based on read depth and quality}
\usage{
FilterSNPs(SNPset, RefAlleleFreq = NULL, FilterAroundMedianDepth = 2.5,
  MinTotalDepth, MaxTotalDepth, MinSampleDepth = NULL, MinGQ = 99,
  Verbose = TRUE)
}
\arguments{
\item{SNPset}{The data frame imported by \code{ImportFromGATK}}

\item{RefAlleleFreq}{A numeric < 1. This will filter out SNPs with a Reference
Allele Frequency less than \code{RefAlleleFreq} and greater than 1 -
\code{RefAlleleFreq}. Eg. \code{RefAlleleFreq = 0.3} will keep SNPs with 0.3 <= REF_FRQ <= 0.7}

\item{FilterAroundMedianDepth}{Filters total SNP read depth for both bulks.
A median and median absolute deviation (MAD) of depth will be calculated. SNPs with
read depth greater or less than \code{FilterAroundMedianDepth} MADs away from the mean
will be filtered.}

\item{MinTotalDepth}{The minimum total read depth for a SNP (counting both bulks)}

\item{MaxTotalDepth}{The maximum total read depth for a SNP (counting both bulks)}

\item{MinSampleDepth}{The minimum read depth for a SNP in each bulk}

\item{MinGQ}{The minimum Genotype Quality as set by GATK. This is a measure of
how confident GATK was with the assigned genotype (i.e. homozygous ref,
heterozygous, homozygous alt). See \href{http://gatkforums.broadinstitute.org/gatk/discussion/1268/what-is-a-vcf-and-how-should-i-interpret-it}{What is a VCF and how should I interpret it?}}

\item{Verbose}{logical. If \code{TRUE} will report number of SNPs filtered in each step.}
}
\value{
Returns a subset of the data frame suplied which meets the filtering
conditions applied by the selected paramaters. If \code{Verbose} is \code{TRUE}
the function reports the number of SNPs filtered in each steps as well as the
initiatl number of SNPs, the total number of SNPs filtered and the remaining number.
}
\description{
A wrapper for the subset function. Use filtering paramaters to filter out high and low depth reads as well as
low Genotype Quality as defined by GATK. All filters are optional but recommended.
}
\examples{
df_filt <- FilterSNPs(
    df,
    RefAlleleFreq = 0.3,
    MinTotalDepth = 40,
    MaxTotalDepth = 80,
    MinSampleDepth = 20,
    MinGQ = 99,
    Verbose = TRUE
)
}
\seealso{
See \code{\link[stats]{mad}} for explaination of calculation of median absolute deviation.
\href{http://gatkforums.broadinstitute.org/gatk/discussion/1268/what-is-a-vcf-and-how-should-i-interpret-it}{What is a VCF and how should I interpret it?}
for more information on GATK Fields and Genotype Fields
}
