% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_functions.R
\name{getFDRThreshold}
\alias{getFDRThreshold}
\title{Find false discovery rate threshold}
\usage{
getFDRThreshold(pvalues, alpha = 0.01)
}
\arguments{
\item{pvalues}{a vector of p-values}

\item{alpha}{the required false discovery rate alpha}
}
\value{
The p-value threshold that corresponds to the Benjamini-Hochberg adjusted p-value at the FDR set by alpha.
}
\description{
Given a vector of p-values and a set false discovery rate alpha the function 
returns the lowest p-value in the vector for which the Benjamini-Hochberg 
adjusted p-value (ie q-value) is less than that alpha.
}
