\name{getG_MH}
\alias{getG_MH}
\title{getG_MH}
\usage{
  getG_MH(LowRef,HighRef,LowAlt,HighAlt)
}
\arguments{
  \item{LowRef}{Low Reference Allele Count}
  
  \item{HighRef}{High Reference Allele Count}
  
  \item{LowAlt}{Low Alternate Allele Count}
  
  \item{HighAlt}{High Alternate Allele Count}
  
}
\description{
  Calculates a G Statistic
}
