\name{getPvals_MH }
\alias{getPvals_MH }
\title{getPvals_MH }
\usage{
  getPvals_MH(Gprime,deltaSNP,outlierFilter,filterThreshold)
}
\arguments{
  \item{Gprime}{Gprime Statistic}
  
  \item{deltaSNP}{deltaSNP}
  
  \item{outlierFilter}{Choose either "deltaSNP" or "Hampel"}
  
  \item{filterThreshold}{Filtering Threshold should be less than 0.5}
  
}
\description{
  Calculates P-values
}
