% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Filter.R
\name{importFromVCF}
\alias{importFromVCF}
\title{importFromVCF}
\usage{
importFromVCF(
  file,
  highBulk = character(),
  lowBulk = character(),
  chromList = NULL,
  filter = FALSE,
  outfile = FALSE
)
}
\arguments{
\item{file}{vcf file}

\item{highBulk}{Highbulk name}

\item{lowBulk}{LowBulk name}

\item{chromList}{chromosome list}

\item{filter}{Boolean Value True or False. If TRUE only variants with "PASS" in the VCF file FILTER column will be used}

\item{outile}{Boolean Value True or False. If TRUE table is written to file.CSV}
}
\value{
Returns a data frame
}
\description{
importFromVCF
}
