% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Filter.R
\name{importFromVCF}
\alias{importFromVCF}
\title{importFromVCF}
\usage{
importFromVCF(
  file,
  highBulk = character(),
  lowBulk = character(),
  chromList = NULL
)
}
\arguments{
\item{file}{vcf file}

\item{highBulk}{Highbulk name}

\item{lowBulk}{LowBulk name}

\item{chromList}{chromosome list}

\item{filter}{Boolean Value True or False. TRUE if you want to filter by "PASS", FALSE you want no filter}

\item{filename}{Provide prefix to file name always ends in .CSV}
}
\value{
Returns a data frame
}
\description{
importFromVCF
}
