% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSAExperimentFromGatkTable.R
\name{makeBSAExperimentFromGatkTable}
\alias{makeBSAExperimentFromGatkTable}
\title{Create a BSAExperiment object from a GATK \code{varianttotable} Table}
\usage{
makeBSAExperimentFromGatkTable(
  gatk_table_path,
  col_data_path,
  drop_samples = c(),
  metadata = list()
)
}
\description{
Create a BSAExperiment object from a GATK \code{varianttotable} Table
}
