% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSAExperimentFromGatkTable.R
\name{makeBSAExperimentFromGatkTable}
\alias{makeBSAExperimentFromGatkTable}
\title{Create a BSAExperiment object from a GATK \code{varianttotable} Table}
\usage{
makeBSAExperimentFromGatkTable(
  gatk_table_path,
  col_data_path,
  drop_samples = c(),
  high_confidence_depth = 10,
  high_confidence_alt_percentage = 0.9,
  keep_multiallelic = FALSE,
  high_confidence_pl = NULL,
  high_confidence_gq = NULL
)
}
\description{
Create a BSAExperiment object from a GATK \code{varianttotable} Table
}
