\name{Obs_Allele_Freq}
\alias{Obs_Allele_Freq}
\title{Obs_Allele_Freq}
\usage{
Obs_Allele_Freq(SNPSet) 
}
\arguments{
\item{SNPSet}{A SNPSet gernerated from a vcf file through the QTL_Parser and importFromTable functions}

}
\description{
 Returns a 4 column data frame with Chromosome, Position, and Observed allele frequencies from the High Parent for both bulks
}
