% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{Obs_Allele_Freq2}
\alias{Obs_Allele_Freq2}
\title{Obs_Allele_Freq2}
\usage{
Obs_Allele_Freq2(SNPSet, ChromosomeValue, threshold)
}
\arguments{
\item{SNPSet}{A SNPSet generated from the function importFromTable}

\item{ChromosomeValue}{Input a Specific Chromosome Value}

\item{threshold}{Input a Specific Allele Frequency Threshold value from the High Bulk High Parent}
}
\description{
Returns a 4 column data filtering for specific Chromosome and High Bulk Observed Allele Frequencies
}
