\name{Obs_Allele_Freq3}
\alias{Obs_Allele_Freq3}
\title{Obs_Allele_Freq3}
\usage{
Obs_Allele_Freq3(SNPSet = df_filt) 
}
\arguments{
\item{SNPSet}{An SNPset generated from importFromTable function and parsed withQTLParser_1_MH}
}
\description{
Generates a data set that can be further analyzed with plotQTLStats_MH
}
