% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{obs_MH}
\alias{obs_MH}
\title{obs_MH}
\usage{
obs_MH(
  SNPSet,
  ChromosomeValue1,
  ChromosomeValue2,
  ChromosomeValue3,
  ChromosomeValue4,
  threshold
)
}
\arguments{
\item{ChromosomeValue1}{Enter a Integer for a specific chromosome i.e. 1,2,..etc.}

\item{ChromosomeValue2}{Enter a Integer for a specific chromosome i.e. 1,2,..etc.}

\item{ChromosomeValue3}{Enter a Integer for a specific chromosome i.e. 1,2,..etc.}

\item{ChromosomeValue4}{Enter a Integer for a specific chromosome i.e. 1,2,..etc.}

\item{threshold}{Specify Alternate Allelic Frequency threshold}

\item{SNPset}{a data frame with SNPs and genotype fields as imported by}
}
\value{
Plots Alternate Allelic Frequency against GPrime value 
  \code{\link[modeest]{mlv}} function from the package modeest. Finally, the
}
\description{
obs_MH
}
\examples{
obs_MH(SNPSet = df_filt, ChromosomeValue1 = 1, ChromosomeValue2 = 2, ChromosomeValue3 = 3, ChromosomeValue4 = 4, threshold = .30)
}
