% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_MH.R
\name{plotGprimeDist_Py_MH}
\alias{plotGprimeDist_Py_MH}
\title{plotGprimeDist_Py_MH}
\usage{
plotGprimeDist_Py_MH(
  SNPset,
  outlierFilter = c("deltaSNP", "Hampel"),
  filterThreshold = 0.1,
  binwidth = 0.5
)
}
\arguments{
\item{SNPset}{An SNPset generated from importFromTable function}

\item{outlierFilter}{Choose outlier filter method either deltaSNP or Hampel}

\item{filterThreshold}{Choose a filter threshold}

\item{binwidth}{Choose binwidth size default is 0.5}
}
\value{
A Gprime Distribution Plot
}
\description{
plotGprimeDist_Py_MH
}
