\name{runGprimeAnalysis_MH}
\alias{runGprimeAnalysis_MH}
\title{runGprimeAnalysis_MH}
\usage{
  runGprimeAnalysis_MH(SNPset, windowSize,outlierFilter,filterThreshold)
}
\arguments{
  \item{SNPset}{A SNP set}
  
  \item{windowSize}{Specify the windowSize}
  
  \item{outlierFilter}{Choose outlier filtering method "deltaSNP" is default value}
  
  \item{filterThreshold}{Choose filtering threshold value default is set to 0.1}
}
\description{
 Runs a Grpime Analysis
}
