% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takagi_sim.R
\name{runQTLseqAnalysis}
\alias{runQTLseqAnalysis}
\title{Calculates delta SNP confidence intervals for QTLseq analysis}
\usage{
runQTLseqAnalysis(SNPset, windowSize = 1e+06, popStruc = "F2", bulkSize,
  depth = NULL, replications = 10000, filter = 0.3, intervals = c(95,
  99), ...)
}
\arguments{
\item{SNPset}{The data frame imported by \code{ImportFromGATK}}

\item{windowSize}{the window size (in base pairs) bracketing each SNP for which to calculate the statitics.}

\item{popStruc}{the population structure. Defaults to "F2" and assumes "RIL" otherwise}

\item{bulkSize}{non-negative integer vector. The number of individuals in
each simulated bulk. Can be of length 1, then both bulks are set to the
same size. Assumes the first value in the vector is the simulated high
bulk.}

\item{depth}{integer. A read depth for which to replicate SNP-index calls.}

\item{replications}{integer. The number of bootstrap replications.}

\item{filter}{numeric. An optional minimum SNP-index filter}

\item{intervals}{numeric vector. Confidence intervals supplied as two-sided
percentiles. i.e. If intervals = '95' will return the two sided 95\%
confidence interval, 2.5\% on each side.}

\item{...}{Other arguments passed to \code{\link[locfit]{locfit}} and
subsequently to \code{\link[locfit]{locfit.raw}}() (or the lfproc). Usefull
in cases where you get "out of vertex space warnings"; Set the maxk higher
than the default 100. See \code{\link[locfit]{locfit.raw}}().  But if you
are getting that warning you should seriously consider increasing your
window size.}
}
\value{
A SNPset data frame with delta SNP-index thresholds corrisponding to the 
requested confidence intervals matching the tricube smoothed depth at each SNP.
}
\description{
The method for simulating delta SNP-index confidence interval thresholds
as described in Takagi et al., (2013). Genotypes are randomly assigned for
each indvidual in the bulk, based on the population structure. The total
alternative allele frequency in each bulk is calculated at each depth used to simulate 
delta SNP-indeces, with a user defined number of bootstrapped replication.
The requested confidence intervals are then calculated from the bootstraps.
}
\examples{
df_filt <- runQTLseqAnalysis(
SNPset = df_filt,
bulkSize = c(25, 35)
windowSize = 1e6,
popStruc = "F2",
replications = 10000,
intervals = c(95, 99)
)

}
