\name{runQTLseqAnalysis_MH}
\alias{runQTLseqAnalysis_MH}
\title{runQTLseqAnalysis_MH}
\usage{
  runQTLseqAnalysis_MH(SNPset,windowSize,popStruc,bulkSize,depth,replications,filter,intervals)
}
\arguments{
  \item{SNPset}{SNPset}
  
  \item{windowSize}{Specify window size}
  
  \item{popStruc}{What is the population strucutre default is F2}
  
  \item{bulkSize}{What are the sizes of the two bulks}
  
  \item{depth}{Specify depth}
  
  \item{replications}{Specify number of replications default is 10000}
  
  \item{filter}{Specify the filtering value defauls is 0.3}
  
  \item{intervals}{Specify the confidence intervals default is 95 - 99}
  
}
\description{
  Run a QTL sequence Analysis
}
